+++ CHANGE_LOG.md
@@ -99,9 +99,13 @@ Change log was not kept before version 2.1.
  * 8.72:
      - Added `CLOG(LEVEL, LoggerId)` macro for sensable name for custom loggers - these macro uses `C##LEVEL` macros
      - `LOG(LEVEL)` now uses newly `CLOG` macros
- * 8.73:
+ * 8.75:
      - Code improvements for speed
      - Added `Loggers::setDefaultConfigurations` to configure existing and future loggers
+ * 8.76:
+     - Introduced `VLOG(verboseLevel)`, `CVLOG(verboseLevel, logger)`, `VLOG_IF`, `VLOG_EVERY_N`, CVLOG_IF`, `CVLOG_EVERY_N` for verbose logging to be more compatible with other major logging libraries
+     - Introduced `VLOG_IS_ON(verboseLevel)` to check if specific verbose level is on
+     - Some code improvements by inlining some code
 
 #### 7.xx
  * 7.00: